requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
timm==0.9.16
pillow==10.3.0 
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2